Move the utilities to planning_scene.cpp
