+++ moveit_core/planning_scene/CMakeLists.txt
@@ -1,6 +1,5 @@
 add_library(moveit_planning_scene SHARED
   src/planning_scene.cpp
-  src/utilities.cpp
 )
 target_include_directories(moveit_planning_scene PUBLIC
   $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
