+++ moveit_core/planning_scene/src/planning_scene.cpp
@@ -36,7 +36,6 @@
 
 #include <boost/algorithm/string.hpp>
 #include <moveit/planning_scene/planning_scene.h>
-#include <moveit/planning_scene/utilities.hpp>
 #include <moveit/collision_detection/occupancy_map.h>
 #include <moveit/collision_detection_fcl/collision_detector_allocator_fcl.h>
 #include <geometric_shapes/shape_operations.h>
@@ -59,6 +58,48 @@ static const rclcpp::Logger LOGGER = rclcpp::get_logger("moveit_planning_scene.p
 
 const std::string PlanningScene::OCTOMAP_NS = "<octomap>";
 const std::string PlanningScene::DEFAULT_SCENE_NAME = "(noname)";
+
+namespace utilities
+{
+/**
+ * convert Pose msg to Eigen::Isometry, normalizing the quaternion part if necessary.
+ * @param msg Input message
+ * @param out Output Eigen transform
+ */
+void poseMsgToEigen(const geometry_msgs::msg::Pose& msg, Eigen::Isometry3d& out)
+{
+  Eigen::Translation3d translation(msg.position.x, msg.position.y, msg.position.z);
+  Eigen::Quaterniond quaternion(msg.orientation.w, msg.orientation.x, msg.orientation.y, msg.orientation.z);
+  quaternion.normalize();
+  out = translation * quaternion;
+}
+
+/** \brief Read a pose from text */
+bool readPoseFromText(std::istream& in, Eigen::Isometry3d& pose)
+{
+  double x, y, z, rx, ry, rz, rw;
+  if (!(in >> x >> y >> z))
+  {
+    RCLCPP_ERROR(LOGGER, "Improperly formatted translation in scene geometry file");
+    return false;
+  }
+  if (!(in >> rx >> ry >> rz >> rw))
+  {
+    RCLCPP_ERROR(LOGGER, "Improperly formatted rotation in scene geometry file");
+    return false;
+  }
+  pose = Eigen::Translation3d(x, y, z) * Eigen::Quaterniond(rw, rx, ry, rz);
+  return true;
+}
+
+/** \brief Write a pose to text */
+void writePoseToText(std::ostream& out, const Eigen::Isometry3d& pose)
+{
+  out << pose.translation().x() << ' ' << pose.translation().y() << ' ' << pose.translation().z() << '\n';
+  Eigen::Quaterniond r(pose.linear());
+  out << r.x() << ' ' << r.y() << ' ' << r.z() << ' ' << r.w() << '\n';
+}
+}  // namespace utilities
 
 class SceneTransforms : public moveit::core::Transforms
 {
@@ -910,7 +951,7 @@ void PlanningScene::saveGeometryToStream(std::ostream& out) const
       {
         out << "* " << id << '\n';  // New object start
         // Write object pose
-        writePoseToText(out, obj->pose_);
+        utilities::writePoseToText(out, obj->pose_);
 
         // Write shapes and shape poses
         out << obj->shapes_.size() << '\n';  // Number of shapes
@@ -918,7 +959,7 @@ void PlanningScene::saveGeometryToStream(std::ostream& out) const
         {
           shapes::saveAsText(obj->shapes_[j].get(), out);
           // shape_poses_ is valid isometry by contract
-          writePoseToText(out, obj->shape_poses_[j]);
+          utilities::writePoseToText(out, obj->shape_poses_[j]);
           if (hasObjectColor(id))
           {
             const std_msgs::msg::ColorRGBA& c = getObjectColor(id);
@@ -932,8 +973,8 @@ void PlanningScene::saveGeometryToStream(std::ostream& out) const
         out << obj->subframe_poses_.size() << '\n';  // Number of subframes
         for (auto& pose_pair : obj->subframe_poses_)
         {
-          out << pose_pair.first << '\n';          // Subframe name
-          writePoseToText(out, pose_pair.second);  // Subframe pose
+          out << pose_pair.first << '\n';                     // Subframe name
+          utilities::writePoseToText(out, pose_pair.second);  // Subframe pose
         }
       }
     }
@@ -993,7 +1034,7 @@ bool PlanningScene::loadGeometryFromStream(std::istream& in, const Eigen::Isomet
 
       // Read in object pose (added in the new scene format)
       pose.setIdentity();
-      if (uses_new_scene_format && !readPoseFromText(in, pose))
+      if (uses_new_scene_format && !utilities::readPoseFromText(in, pose))
       {
         RCLCPP_ERROR(LOGGER, "Failed to read object pose from scene file");
         return false;
@@ -1012,7 +1053,7 @@ bool PlanningScene::loadGeometryFromStream(std::istream& in, const Eigen::Isomet
           RCLCPP_ERROR(LOGGER, "Failed to load shape from scene file");
           return false;
         }
-        if (!readPoseFromText(in, pose))
+        if (!utilities::readPoseFromText(in, pose))
         {
           RCLCPP_ERROR(LOGGER, "Failed to read pose from scene file");
           return false;
@@ -1048,7 +1089,7 @@ bool PlanningScene::loadGeometryFromStream(std::istream& in, const Eigen::Isomet
         {
           std::string subframe_name;
           in >> subframe_name;
-          if (!readPoseFromText(in, pose))
+          if (!utilities::readPoseFromText(in, pose))
           {
             RCLCPP_ERROR(LOGGER, "Failed to read subframe pose from scene file");
             return false;
@@ -1348,7 +1389,7 @@ void PlanningScene::processOctomapMsg(const octomap_msgs::msg::OctomapWithPose& 
 
   const Eigen::Isometry3d& t = getFrameTransform(map.header.frame_id);
   Eigen::Isometry3d p;
-  poseMsgToEigen(map.origin, p);
+  utilities::poseMsgToEigen(map.origin, p);
   p = t * p;
   world_->addToObject(OCTOMAP_NS, std::make_shared<const shapes::OcTree>(om), p);
 }
@@ -1466,7 +1507,7 @@ bool PlanningScene::processAttachedCollisionObjectMsg(const moveit_msgs::msg::At
         Eigen::Isometry3d subframe_pose;
         for (std::size_t i = 0; i < object.object.subframe_poses.size(); ++i)
         {
-          poseMsgToEigen(object.object.subframe_poses[i], subframe_pose);
+          utilities::poseMsgToEigen(object.object.subframe_poses[i], subframe_pose);
           std::string name = object.object.subframe_names[i];
           subframe_poses[name] = subframe_pose;
         }
@@ -1667,7 +1708,7 @@ bool PlanningScene::shapesAndPosesFromCollisionObjectMessage(const moveit_msgs::
   shapes.reserve(num_shapes);
   shape_poses.reserve(num_shapes);
 
-  poseMsgToEigen(object.pose, object_pose);
+  utilities::poseMsgToEigen(object.pose, object_pose);
 
   bool switch_object_pose_and_shape_pose = false;
   if (num_shapes == 1)
@@ -1684,7 +1725,7 @@ bool PlanningScene::shapesAndPosesFromCollisionObjectMessage(const moveit_msgs::
     if (!s)
       return;
     Eigen::Isometry3d pose;
-    poseMsgToEigen(pose_msg, pose);
+    utilities::poseMsgToEigen(pose_msg, pose);
     if (!switch_object_pose_and_shape_pose)
     {
       shape_poses.emplace_back(std::move(pose));
@@ -1765,7 +1806,7 @@ bool PlanningScene::processCollisionObjectAdd(const moveit_msgs::msg::CollisionO
   Eigen::Isometry3d subframe_pose;
   for (std::size_t i = 0; i < object.subframe_poses.size(); ++i)
   {
-    poseMsgToEigen(object.subframe_poses[i], subframe_pose);
+    utilities::poseMsgToEigen(object.subframe_poses[i], subframe_pose);
     std::string name = object.subframe_names[i];
     subframes[name] = subframe_pose;
   }
@@ -1807,7 +1848,7 @@ bool PlanningScene::processCollisionObjectMove(const moveit_msgs::msg::Collision
     const Eigen::Isometry3d& world_to_object_header_transform = getFrameTransform(object.header.frame_id);
     Eigen::Isometry3d header_to_pose_transform;
 
-    poseMsgToEigen(object.pose, header_to_pose_transform);
+    utilities::poseMsgToEigen(object.pose, header_to_pose_transform);
 
     const Eigen::Isometry3d object_frame_transform = world_to_object_header_transform * header_to_pose_transform;
     world_->setObjectPose(object.id, object_frame_transform);
